Update virtualcenter, check vm exists on status
There's a code path that leads to ObjectNotFound exceptions being raised
If provider.is_vm_[running|stopped|suspended] is called on a VM that
doesn't exist, the call to _get_vm ends up raising an ugly exception.

Check that the VM exists before calling _get_vm on it in vm_status.