Default entities for Nuage provider
With this commit we implement all the code needed to be
able to manage a specific enterprise (organization) on Nuage
server. Specifically, we support creation of following test
entities:

- enterprise with default name
- domain template with defailt name
- domain with default name
- zone with default name
- subnet with default name

Also, we support deleteion of all these entities. Why we need this?
We have an integration test that checks whether appropriate events
are emitted, and we need to be able to create some isolated set of
entities to trigger events on.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>