Add Redfish chassis and rack related functionality
The Redfish API now supports retrieving Chassis resources, which
are general chassis entities. It also provides helper methods for
retrieving physical racks, which are a special type of Chassis.
The Redfish racks currently support getting the name of the rack.

For physical chassis, we are able to obtain number of servers
installed within a chassis, the name and description of the cassis,
and the status of identifying LED, if any.

This functionality is related to the following Bugzilla items:
  * https://bugzilla.redhat.com/show_bug.cgi?id=1571610
  * https://bugzilla.redhat.com/show_bug.cgi?id=1574809