[RFR] Adds Redfish API client (#328)
* add Redfish API client

The Redfish provider's implementation now uses the Redfish API
client [1].

[1] https://pypi.org/project/redfish-client/

* Add new entity representing servers

Server is a base entity that exists side-by-side with VM-like
entities. It represents persistent, possibly physical servers
(e.g., bare metal).

This patch provides a general implementation of the classes
and methods that describe servers and their states.

* Implement stats for Redfish physical servers

This adds the handling of the stat retrieval from the Redfish
API. We implemented the stats that are currently supported at
the provider end.

The implementation of the API stat retrieval is a specialisation
of the `Server` entity, providing an example for possible other
future (or past) (physical) servers.