[RFR] Add API for Redfish chassis navigation (#335)
* Add counts of Redfish chassis and racks

The Redfish API now retrieves the number of Chassis and Racks.

Related Bugzilla ticket:
  * https://bugzilla.redhat.com/show_bug.cgi?id=1571610

* Abstract Redfish resource class

In Redfish, we normally get top-level resources `Systems`,
`Chassis`, `EventService` and possible others. They share
certain properties and methods, therefore it will be an
easier way forward if the existing `RedfishServer` had
an ancester in the more general new `RedfishResource`.

* Add new entity Physical Container

In the context of physical infrastructure, servers normally
reside in one or more nested enclosures such as chassis or racks.
This new entity serves as a reporesentation of these physical
containers.
Not to be confused by container virtualisation technology.

* Add Redfish chassis and rack related functionality

The Redfish API now supports retrieving Chassis resources, which
are general chassis entities. It also provides helper methods for
retrieving physical racks, which are a special type of Chassis.
The Redfish racks currently support getting the name of the rack.

For physical chassis, we are able to obtain number of servers
installed within a chassis, the name and description of the cassis,
and the status of identifying LED, if any.

This functionality is related to the following Bugzilla items:
  * https://bugzilla.redhat.com/show_bug.cgi?id=1571610
  * https://bugzilla.redhat.com/show_bug.cgi?id=1574809

* Backport py2+py3 compatibility of ABCMeta classes

We used `six.with_metaclass` to describe all the entity classes
for compatibility of Python 2 and Python 3.