Modify the Redfish resources' name
The name property of the Redfish resources such as Physical Systems
or Chassis other than Racks now reflect the method of composing the
name in the ManageIQ's GUI: concatenation of Manufacturer, Name
and, if available, Serial Number in parentheses.

The name of Rack instances is unchanged, i.e., it shows the Id of
the resource. To distinguish the representation from other types
of physical chassis in Redfish, we introduced the `RedfishRack`
class.

Related PR:
  * https://github.com/ManageIQ/manageiq-providers-redfish/pull/61